requirements.txt
Pillow
pywin32
tkinterdnd2